pyproject.toml
[project]
name = "flask-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask<3",
    "python-dotenv>=1.2.2",
    "werkzeug<3",
]
